backend/requirements.txt
fastapi
uvicorn
yfinance
redis
pydantic
pydantic-settings
python-dotenv
httpx
numpy
pandas
scipy
torch
scikit-learn
xgboost
arch
vaderSentiment
joblib
ta
sse-starlette
websockets
sqlalchemy
asyncpg
alembic
apscheduler
aiosqlite
greenlet
PyJWT
supabase